# jag_api_contract_v1.yaml
openapi: 3.1.0
info:
  title: JAG Platform API
  version: 1.0.0
  description: |
    JAG Holdings unified platform API — Core, IMS, JABCO PM, CRM, Entertainment, Family, and Properties modules.

    **Authentication:** Keycloak bearer JWT required on all endpoints except `GET /health`.
    Tokens are issued by the Keycloak realm; no custom auth endpoints exist in this API.

    **Response envelope (STD-06):** Every response uses `{ success, data, error, code, meta? }`.
    Errors never expose raw stack traces.

    **Idempotency (STD-11):** All financial write endpoints require `Idempotency-Key: <uuid>`.
    A duplicate key returns the first result without re-execution; response status 409.

    **Versioning (STD-05):** All endpoints are prefixed `/api/v1/`. Breaking changes → `/api/v2/`.

    **Pagination:** All list endpoints accept `?page=&limit=&sort=&order=` and return
    `meta: { total, page, limit, pages }`.

    ---
    ### PRE-2 Design Decisions

    1. **pending_review_queue gap** — flagged in PRE-1, not yet in schema. Failed/disputed WiPay
       webhook events currently return 422 `WIPAY_VALIDATION_FAILED`. If a `pending_review_queue`
       table is added (jag_core or jag_properties), the webhook handler should INSERT a row and
       return 202 instead. Raise in PRE-3/PRE-4.

    2. **entity_tag** — `BAR | MEMBERS_CLUB` is a required field on every Entertainment
       transaction request body and response. Non-negotiable per PRE-1 design decision 4.

...

...

...
servers:
...
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Keycloak-issued JWT. Obtain from the JAG Keycloak realm token endpoint.
...
  parameters:
    UuidPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortParam:
      name: sort
      in: query
      schema:
        type: string
    OrderParam:
      name: order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  schemas:
    PaginationMeta:
      type: object
      required: [total, page, limit, pages]
      properties:
        total:
          type: integer
          example: 42
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        pages:
          type: integer
          example: 3
    ApiResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
        data:
          description: Response payload — schema varies per endpoint. Null on error.
          nullable: true
        meta:
          $ref: '#/components/schemas/PaginationMeta'
        error:
          type: string
          nullable: true
          description: Human-readable error message. Present only on failure.
        code:
          type: string
          nullable: true
          description: >
            Machine-readable error code. Present only on failure.
            Examples: VALIDATION_ERROR, UNAUTHORIZED, FORBIDDEN, NOT_FOUND,
            IDEMPOTENCY_CONFLICT, BUSINESS_RULE_VIOLATION, INTERNAL_ERROR.

    # -------------------------------------------------------
    # CORE SCHEMAS
    # -------------------------------------------------------
...
    PropProperty:
      type: object
      properties:
        id:
          type: string
          format: uuid
        owner_id:
          type: string
          format: uuid
        property_code:
          type: string
          example: PROP-BARATARIA-01
        name:
          type: string
          example: Barataria House 1
        address_line1:
          type: string
        address_line2:
          type: string
          nullable: true
        city:
          type: string
        region:
          type: string
          nullable: true
        country:
          type: string
          example: TT
        property_type:
          type: string
          enum: [RESIDENTIAL, COMMERCIAL, LAND, MIXED, AGRICULTURAL]
        tenure_type:
          type: string
          enum: [FREEHOLD, LEASEHOLD, STATE_LAND]
        lot_number:
          type: string
          nullable: true
        lot_size_sqm:
          type: number
          nullable: true
        floor_area_sqm:
          type: number
          nullable: true
        bedrooms:
          type: integer
          nullable: true
        bathrooms:
          type: integer
          nullable: true
        purchase_date:
          type: string
          format: date
          nullable: true
        purchase_price:
          type: number
          nullable: true
        current_valuation:
          type: number
          nullable: true
        valuation_date:
          type: string
          format: date
          nullable: true
        is_rented:
          type: boolean
        is_active:
          type: boolean
        notes:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    PropPropertyCreate:
      type: object
      required: [property_code, name, address_line1, city, property_type, tenure_type]
      properties:
        property_code:
          type: string
        name:
          type: string
        address_line1:
          type: string
        address_line2:
          type: string
          nullable: true
        city:
          type: string
        region:
          type: string
          nullable: true
        country:
          type: string
          default: TT
        property_type:
          type: string
          enum: [RESIDENTIAL, COMMERCIAL, LAND, MIXED, AGRICULTURAL]
        tenure_type:
          type: string
          enum: [FREEHOLD, LEASEHOLD, STATE_LAND]
          default: FREEHOLD
        lot_number:
          type: string
          nullable: true
        lot_size_sqm:
          type: number
          nullable: true
        floor_area_sqm:
          type: number
          nullable: true
        bedrooms:
          type: integer
          nullable: true
        bathrooms:
          type: integer
          nullable: true
        purchase_date:
          type: string
          format: date
          nullable: true
        purchase_price:
          type: number
          nullable: true
        notes:
          type: string
          nullable: true
    PropPropertyUpdate:
      type: object
      properties:
        name:
          type: string
        current_valuation:
          type: number
          nullable: true
        valuation_date:
          type: string
          format: date
          nullable: true
        is_rented:
          type: boolean
        is_active:
          type: boolean
        notes:
          type: string
          nullable: true
...
    PropMaintenanceRequest:
      type: object
      properties:
        id:
          type: string
          format: uuid
        owner_id:
          type: string
          format: uuid
        property_id:
          type: string
          format: uuid
        lease_id:
          type: string
          format: uuid
          nullable: true
        reported_by_tenant_id:
          type: string
          format: uuid
          nullable: true
        category:
          type: string
          enum: [PLUMBING, ELECTRICAL, STRUCTURAL, HVAC, APPLIANCE, PEST_CONTROL, SECURITY, GARDEN, PAINTING, ROOFING, OTHER]
        description:
          type: string
        priority:
          type: string
          enum: [LOW, MEDIUM, HIGH, URGENT]
        status:
          type: string
          enum: [OPEN, ASSIGNED, IN_PROGRESS, AWAITING_PARTS, COMPLETED, CLOSED, CANNOT_REPRODUCE]
        assigned_to:
          type: string
          nullable: true
          description: Contractor name or internal reference
        estimated_cost:
          type: number
          nullable: true
        actual_cost:
          type: number
          nullable: true
        reported_date:
          type: string
          format: date
        scheduled_date:
          type: string
          format: date
          nullable: true
        completed_date:
          type: string
          format: date
          nullable: true
        completion_notes:
          type: string
          nullable: true
        idempotency_key:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    PropMaintenanceRequestCreate:
      type: object
      required: [property_id, category, description, reported_date]
      properties:
        property_id:
          type: string
          format: uuid
        lease_id:
          type: string
          format: uuid
          nullable: true
        reported_by_tenant_id:
          type: string
          format: uuid
          nullable: true
        category:
          type: string
          enum: [PLUMBING, ELECTRICAL, STRUCTURAL, HVAC, APPLIANCE, PEST_CONTROL, SECURITY, GARDEN, PAINTING, ROOFING, OTHER]
        description:
          type: string
        priority:
          type: string
          enum: [LOW, MEDIUM, HIGH, URGENT]
          default: MEDIUM
        reported_date:
          type: string
          format: date
        estimated_cost:
          type: number
          nullable: true
    PropMaintenanceRequestUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [OPEN, ASSIGNED, IN_PROGRESS, AWAITING_PARTS, COMPLETED, CLOSED, CANNOT_REPRODUCE]
        priority:
          type: string
          enum: [LOW, MEDIUM, HIGH, URGENT]
        assigned_to:
          type: string
          nullable: true
        estimated_cost:
          type: number
          nullable: true
        actual_cost:
          type: number
          nullable: true
        scheduled_date:
          type: string
          format: date
          nullable: true
        completed_date:
          type: string
          format: date
          nullable: true
        completion_notes:
          type: string
          nullable: true
...
    FamVehicleMaintenance:
      type: object
      properties:
        id:
          type: string
          format: uuid
        owner_id:
          type: string
          format: uuid
        vehicle_id:
          type: string
          format: uuid
        service_type:
          type: string
          enum: [OIL_CHANGE, TYRE_ROTATION, BRAKE_SERVICE, ANNUAL_SERVICE, TRANSMISSION_SERVICE, AC_SERVICE, WINDSCREEN, BODY_REPAIR, INSPECTION, REPAIR, OTHER]
        service_date:
          type: string
          format: date
        mileage_at_service:
          type: integer
          nullable: true
        service_provider:
          type: string
          nullable: true
        cost:
          type: number
          nullable: true
        description:
          type: string
          nullable: true
        next_service_date:
          type: string
          format: date
          nullable: true
        next_service_mileage:
          type: integer
          nullable: true
        receipt_path:
          type: string
          nullable: true
          description: MinIO path for receipt/invoice
        idempotency_key:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    FamVehicleMaintenanceCreate:
      type: object
      required: [service_type, service_date]
      properties:
        service_type:
          type: string
          enum: [OIL_CHANGE, TYRE_ROTATION, BRAKE_SERVICE, ANNUAL_SERVICE, TRANSMISSION_SERVICE, AC_SERVICE, WINDSCREEN, BODY_REPAIR, INSPECTION, REPAIR, OTHER]
        service_date:
          type: string
          format: date
        mileage_at_service:
          type: integer
          nullable: true
        service_provider:
          type: string
          nullable: true
        cost:
          type: number
          nullable: true
        description:
          type: string
          nullable: true
        next_service_date:
          type: string
          format: date
          nullable: true
        next_service_mileage:
          type: integer
          nullable: true
        receipt_path:
          type: string
          nullable: true
...
    ErrorResponse:
      type: object
      required: [success, data, error, code]
      properties:
        success:
          type: boolean
          enum: [false]
        data:
          nullable: true
          default: null
        error:
          type: string
          description: Human-readable error message.
        code:
          type: string
          description: Machine-readable error code.
  responses:
    BadRequest:
      description: Validation error (STD-10 — Zod validation failure)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            data: null
            error: "Field 'email' must be a valid email address."
            code: VALIDATION_ERROR
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            data: null
            error: Authentication required.
            code: UNAUTHORIZED
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            data: null
            error: Resource not found.
            code: NOT_FOUND
    Conflict:
      description: Idempotency key already used — returns original result (STD-11)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            data: null
            error: This Idempotency-Key has already been used.
            code: IDEMPOTENCY_CONFLICT
...
paths:
...
  /family/vehicles/{id}/maintenance:
    parameters:
      - $ref: '#/components/parameters/UuidPath'
    get:
      tags: [Family]
      summary: List maintenance history for a vehicle
      operationId: listFamVehicleMaintenance
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SortParam'
        - $ref: '#/components/parameters/OrderParam'
      responses:
        '200':
          description: Maintenance record list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/FamVehicleMaintenance'
                      meta:
                        $ref: '#/components/schemas/PaginationMeta'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [Family]
      summary: Log a maintenance service entry
      description: Financial write — `Idempotency-Key` required (STD-11).
      operationId: createFamVehicleMaintenance
      parameters:
        - in: header
          name: Idempotency-Key
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FamVehicleMaintenanceCreate'
      responses:
        '201':
          description: Maintenance entry logged
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/FamVehicleMaintenance'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /properties:
    get:
      tags: [Properties]
      summary: List properties
      operationId: listProperties
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SortParam'
        - $ref: '#/components/parameters/OrderParam'
        - name: property_type
          in: query
          schema:
            type: string
            enum: [RESIDENTIAL, COMMERCIAL, LAND, MIXED, AGRICULTURAL]
        - name: is_rented
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Property list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/PropProperty'
                      meta:
                        $ref: '#/components/schemas/PaginationMeta'
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Properties]
      summary: Add a property
      operationId: createProperty
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PropPropertyCreate'
      responses:
        '201':
          description: Property created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PropProperty'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /properties/{id}:
    parameters:
      - $ref: '#/components/parameters/UuidPath'
    get:
      tags: [Properties]
      summary: Get property by ID
      operationId: getProperty
      responses:
        '200':
          description: Property record
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PropProperty'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Properties]
      summary: Update property
      operationId: updateProperty
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PropPropertyUpdate'
      responses:
        '200':
          description: Property updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PropProperty'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /properties/maintenance:
    get:
      tags: [Properties]
      summary: List maintenance requests
      operationId: listMaintenanceRequests
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SortParam'
        - $ref: '#/components/parameters/OrderParam'
        - name: status
          in: query
          schema:
            type: string
            enum: [OPEN, ASSIGNED, IN_PROGRESS, AWAITING_PARTS, COMPLETED, CLOSED, CANNOT_REPRODUCE]
        - name: priority
          in: query
          schema:
            type: string
            enum: [LOW, MEDIUM, HIGH, URGENT]
        - name: property_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Maintenance request list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/PropMaintenanceRequest'
                      meta:
                        $ref: '#/components/schemas/PaginationMeta'
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Properties]
      summary: Create maintenance request
      operationId: createMaintenanceRequest
      parameters:
        - in: header
          name: Idempotency-Key
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PropMaintenanceRequestCreate'
      responses:
        '201':
          description: Maintenance request created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PropMaintenanceRequest'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /properties/maintenance/{id}:
    parameters:
      - $ref: '#/components/parameters/UuidPath'
    get:
      tags: [Properties]
      summary: Get maintenance request by ID
      operationId: getMaintenanceRequest
      responses:
        '200':
          description: Maintenance request record
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PropMaintenanceRequest'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Properties]
      summary: Update maintenance request
      operationId: updateMaintenanceRequest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PropMaintenanceRequestUpdate'
      responses:
        '200':
          description: Maintenance request updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PropMaintenanceRequest'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...